+++ src/components/CalloutToForum.tsx
@@ -1,18 +1,14 @@
 'use client'
-import Link from 'next/link'
 import { Callout } from './Callout'
 
 type Props = {}
 
 export function CalloutToForum(_props: Props) {
   return (
-    <Callout title="Get help">
+    <Callout title="문의하기">
       <div className="mb-4">
-        Got any questions about this page? Feel free to ask them on{' '}
-        <Link href={`https://forum.speedetail.app/c/help/19`} target="_blank">
-          the user forum here
-        </Link>
-        .
+        궁금하신 점이 있나요? <a href="mailto:speedetail2@gmail.com">이메일</a>
+        로 편하게 문의해 주세요.
       </div>
     </Callout>
   )