+++ src/lib/redirects.ts
@@ -5,10 +5,10 @@ export const Redirects = () => {
   useEffect(() => {
     const { pathname, hash } = window.location
     if (
-      pathname === '/reference/main-user-interface' &&
+      pathname === '/app-version/main-user-interface' &&
       hash === '#user-data-directory'
     ) {
-      window.location.href = '/reference/user-data-directory'
+      window.location.href = '/app-version/main-user-interface'
     }
   }, [])
   return null