+++ src/markdoc/search.mjs
@@ -29,7 +29,8 @@ function extractSections(node, sections, isRoot = true) {
   if (node.type === 'heading' || node.type === 'paragraph') {
     let content = toString(node).trim()
     if (node.type === 'heading' && node.attributes.level <= 2) {
-      let hash = node.attributes?.id ?? slugify(content)
+      // let hash = node.attributes?.id ?? slugify(content)
+      let hash = node.attributes?.id ?? content
       sections.push([content, hash, []])
     } else {
       sections.at(-1)[2].push(content)